feed/write page에서 ''use client' 쓰지 않도록
- page에서 use client를 쓰는건 자제해야됌...? seo를 할 수 없고 ssr의 장점을 가질 수 없으니까..?
